+++ lib/src/block_parser.dart
@@ -4,25 +4,24 @@
 
 import 'ast.dart';
 import 'block_syntaxes/block_syntax.dart';
-// import 'block_syntaxes/block_tag_block_html_syntax.dart';
 import 'block_syntaxes/blockquote_syntax.dart';
 import 'block_syntaxes/code_block_syntax.dart';
 import 'block_syntaxes/dummy_block_syntax.dart';
 import 'block_syntaxes/empty_block_syntax.dart';
 import 'block_syntaxes/header_syntax.dart';
 import 'block_syntaxes/horizontal_rule_syntax.dart';
-// import 'block_syntaxes/long_block_html_syntax.dart';
+import 'block_syntaxes/html_block_syntax.dart';
 import 'block_syntaxes/ordered_list_syntax.dart';
-// import 'block_syntaxes/other_tag_block_html_syntax.dart';
 import 'block_syntaxes/paragraph_syntax.dart';
 import 'block_syntaxes/setext_header_syntax.dart';
 import 'block_syntaxes/unordered_list_syntax.dart';
 import 'document.dart';
+import 'line.dart';
 
 /// Maintains the internal state needed to parse a series of lines into blocks
 /// of Markdown suitable for further inline parsing.
 class BlockParser {
-  final List<String> lines;
+  final List<Line> lines;
 
   /// The Markdown document this parser is parsing.
   final Document document;
@@ -43,15 +42,7 @@ class BlockParser {
   /// The collection of built-in block parsers.
   final List<BlockSyntax> standardBlockSyntaxes = [
     const EmptyBlockSyntax(),
-    // const BlockTagBlockHtmlSyntax(),
-    // LongBlockHtmlSyntax(r'^ {0,3}<pre(?:\s|>|$)', '</pre>'),
-    // LongBlockHtmlSyntax(r'^ {0,3}<script(?:\s|>|$)', '</script>'),
-    // LongBlockHtmlSyntax(r'^ {0,3}<style(?:\s|>|$)', '</style>'),
-    // LongBlockHtmlSyntax('^ {0,3}<!--', '-->'),
-    // LongBlockHtmlSyntax(r'^ {0,3}<\?', r'\?>'),
-    // LongBlockHtmlSyntax('^ {0,3}<![A-Z]', '>'),
-    // LongBlockHtmlSyntax(r'^ {0,3}<!\[CDATA\[', r'\]\]>'),
-    // const OtherTagBlockHtmlSyntax(),
+    const HtmlBlockSyntax(),
     const SetextHeaderSyntax(),
     const HeaderSyntax(),
     const CodeBlockSyntax(),
@@ -73,10 +64,10 @@ class BlockParser {
   }
 
   /// Gets the current line.
-  String get current => lines[_pos];
+  Line get current => lines[_pos];
 
   /// Gets the line after the current one or `null` if there is none.
-  String? get next {
+  Line? get next {
     // Don't read past the end.
     if (_pos >= lines.length - 1) return null;
     return lines[_pos + 1];
@@ -88,7 +79,7 @@ class BlockParser {
   /// `peek(0)` is equivalent to [current].
   ///
   /// `peek(1)` is equivalent to [next].
-  String? peek(int linesAhead) {
+  Line? peek(int linesAhead) {
     if (linesAhead < 0) {
       throw ArgumentError('Invalid linesAhead: $linesAhead; must be >= 0.');
     }
@@ -101,27 +92,55 @@ class BlockParser {
     _pos++;
   }
 
+  void retreat() {
+    _pos--;
+  }
+
   bool get isDone => _pos >= lines.length;
 
   /// Gets whether or not the current line matches the given pattern.
   bool matches(RegExp regex) {
     if (isDone) return false;
-    return regex.hasMatch(current);
+    return regex.hasMatch(current.content);
   }
 
   /// Gets whether or not the next line matches the given pattern.
   bool matchesNext(RegExp regex) {
     if (next == null) return false;
-    return regex.hasMatch(next!);
+    return regex.hasMatch(next!.content);
   }
 
-  List<Node> parseLines() {
+  /// The parent [BlockSyntax] when it is running inside a nested syntax.
+  BlockSyntax? get parentSyntax => _parentSyntax;
+  BlockSyntax? _parentSyntax;
+
+  List<Node> parseLines({
+    BlockSyntax? parentSyntax,
+  }) {
+    _parentSyntax = parentSyntax;
     final blocks = <Node>[];
+
+    // If the `_pos` does not change before and after `parse()`, never try to
+    // parse the line at `_pos` with the same syntax again.
+    // For example the `TableSyntax` might not advance the `_pos` in `parse`
+    // method, beause of the header row does not match the delimiter row in the
+    // number of cells, which makes a table like structure not be recognized.
+    BlockSyntax? neverMatch;
+
     while (!isDone) {
       for (final syntax in blockSyntaxes) {
+        if (neverMatch == syntax) {
+          continue;
+        }
+
         if (syntax.canParse(this)) {
+          final positionBefore = _pos;
           final block = syntax.parse(this);
-          if (block != null) blocks.add(block);
+          if (block != null) {
+            blocks.add(block);
+          }
+          neverMatch = _pos != positionBefore ? null : syntax;
+
           break;
         }
       }