+++ test/block_syntax/fenced_tex_block_syntax_test.dart
@@ -20,7 +20,7 @@ $a$
 ''';
 
     /// 문장 쪼개기
-    final lines = LineSplitter.split(data).toList();
+    final lines = LineSplitter.split(data).cast<Line>().toList();
 
     /// 파싱하기
     final nodes = BlockParser(lines, document).parseLines();